Introduce check_multiple_package_version actor
Upgrade should be inhibited if both x32 and x64 version of certain packages
are simultaneously present on the machine.
The actor will check for those packages and produce a Report with a bugzilla
link (if any is present for the problematic package).

The implementation also modifies the create_lookup library function
allowing it to create multi-parameter lookups. Regarding the `context`
parameter is supposed to be used changed just for the testing purposes,
the new `arch` parameter is put before the `context` one to make
use of the function more friendly.

OAMG-4097