repomapping: ignore mapping for the extras repositories
Regarding the upgrade for SAP HANA, we want to map only E4S to E4S
repositories. However, Extras repository is not E4S (it's one repo
for whole major release) and we cannot map it just to particular
repository (standard OR EUS OR E4S ...) only. So let's ignore it.
As well, do not report the Extras repository as unknown to upgrade.