Check the FSTAB file for the deprecated / removed XFS mount options.
Some mount options for XFS have been deprecated on RHEL 7 and already
removed on RHEL 8. If any such an option is present in the FSTAB,
it's impossible to boot the RHEL 8 system without the manual update
of the file.

Check whether any of these options are present in the FSTAB file
and inhibit the upgrade in such a case.