s390 inhibit if /boot is not on a separate partition
Previously, when the upgrade is performed on a s390x machine with zipl
and /boot is not on a separate partition, the content of /boot got
deleted during the upgrade process. Since the potential solution for
this problem turned out to fail the upgrade for all s390 upgrades, it
has been decided to create an actor that inhibits the upgrade for the
detected scenario that caused the deletion of the /boot content.

This patch introduces a new check actor that inhibits the upgrade when:
- The architecture is s390X
- /boot is not a mounted partition

In case all of the above is true, the upgrade is inhibited with a
report.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>