upgradeinitramfsgenerator: Handle correctly copy of dracut modules
In case multiple tasks for the same dracut modules exists, the actor
crashed with the traceback. In case the destination path exists
(a module has been copied already) skip the task and log debug
message.

The copy_dracut_modules function was not covered by tests before.
Improve the coverage by unit tests for this function to cover the
discovered problem.