peseventsscanner: fix pylint (C0103) and invalid unit-tests
Renamed attributes of classes inside the actor to the UPPER_CASE
format to make pylint happy.

As well, it was discovered that unit-tests are broken - some unit tests
are not run because of invalid name of function and such tests contains
invalid code as well. This is fixed as well.