amultiple attempts to fetch data from server and increased timeout
The original timout for the fetching of data was too low (just 5s).
Keep 5s for connection to the server and at least 30 seconds to read
the data.

As well, try to fetch data multiple times to reduce problems with
random network glitches. Currently 3 attempts per fetch of a data
file.

In case the reading of data is too slow (no connection timeout)
increase the timeout by 10s per attempt.