Use full module path for requests exceptions, fix tuple use
Most exceptions in requests.exceptions can be also imported from the
parent module. However, in requests 2.7.0 (the version on RHEL 7),
requests.ReadTimeout is only available as requests.exceptions.ReadTimeout
and attempting to access it the first way leads to a traceback.
For consistency, other exceptions are used with full paths too.