Update the RestrictedPCIDevice model and doc changes
The RestrictedPCIdevice model is not designed well for a generic use.
Currently all actors working with the model are usable just for one
specific repository and there is not a nice way to work with the data
in a generic way as expected - mainly because the unsupported_rhelX
and available_rhelX fields, where X in {7, 8, 9}.

Providing two new fields *unsupported* and *available* which contains
list of related RHEL major versions. Actors could work with those
data based on the used target system without any magic.