xfsinfoscanner: Do not use systemd-mount data
The data from systemd-mount is not relevant for the xfsinfoscanner. All
the parititions mounted are reported through /proc/mounts and the output
as is from systemd-mount --list is not giving us any additionally useful
output. It rather caused problems by having PCI addresses as mount
paths, which caused failures when we tried to get the information if the
xfs partition has ftype=0.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>