Introduce the el8toel9 repository + extend system_upgrade/common repo
Move stuff below from the el7toel8 repository to the system_upgrade/common
repository:
  - IPUWorkflow
  - CPUInfo model
  - scancpu actor
All above is used by the el8toel9 repository too.

The el8toel9 repository depends on common and system_upgrade/common
repositories. The basic directory structure for the repository
is created as well for the sake of calm mind, so people are prevented
from troubles made by a typo in any of these directories name.

We expect we do not need any new workflow for the el8toel9 repository
in future, so I am skipping this one.

As well for the benefits, we add the bundled-rpms directory into
the files dir as we have already customers & developers that are
using that directory to tell that specific local RPMs should be
installed during the upgrade. As well, we are using it for the
installation of deps packages during the upgrade. To keep the same
experience, it's there.

Currently to have "the empty" dirs, we put inside the .gitkeep files
that are empty and are expected to be removed in future when a real
content is put inside.