Packaging: support building on RHEL 8
Add support for building on RHEL 8 systems for IPU 8 -> 9.
Packages on RHEL X should contain just stuff related for the
particular system, a la IPU X -> X+1. So keep the el7toel8 repo just
on RHEL 7 system and keep the el8toel9 repo just on RHEL 8.

The dependencies has been updated to reflect the system version.
Names of deps packages that are bundled into the repositories for
the later installation during the IPU are build for both IPU now.
In one way it will be *-el8 packages, for the new IPU it will be
*-el9 packages. All copied into the expected leapp repositories.

This relates to the change in the Makefile where the building of
deps subpackages is executed two times:
  - once for RHEL 8
  - once for RHEL 9
all built deps packages are part of the second source. Additionally
when the CHROOT_COPR is not specified, do not set by default the
rhel-7-x86_64 CHROOT, so build on all available chroots.

The SPEC files has been a little bit polished, but to be hones, I do
not have idea how to make the more readable like in case of Leapp.
In this case I think this is best we can do for now.