config.version: Fix detection of the RHEL-ALT
The is_rhel_alt() function from the config.version library is valid
only on RHEL 7 systems as the detection is based on the major version
of booted kernel:
    RHEL 7     :  3.x
    RHEL-ALT 7 :  4.x

But RHEL 8+ systems have kernel 4+, so the function returned True
on all RHEL 8 systems. Added check to the function - returns False
for all non RHEL 7 systems always.