mounting.py: update systemd-nspawn option on RHEL 9+
Set the --keep-unit and --capability=all options on rhel8+ for
systemd-nspawn. It's typical that we do not have dbus inside the
container, however since RHEL 9, we are not able to create another
container from a container without the --keep-unit option which
is expected to be used in such a case.

Currently we affect only RHEL 9+ systems until we check it's ok to
use it on all systems (we want to keep IPU 7 -> 8 for now to save
a possible debugging of issues in future)