Tests: improve search of an ACTOR: support multiple leapp repos
With the introduced common & el8toel9 repositories, the script searching
the ACTOR returns only the first discovered ACTOR of the given name.
However, it is expected that actors of the same name could exists
in el7toel8 and el8toel9 repositories as these two are independent
on each other. To fix this, we use for now a private method
_lookup_actors() introduced by:
     https://github.com/oamg/leapp/pull/722
Currently it's not possible to specify which repository should be
tested. Right now all discovered actors of the given name are
going to be tested when discovered.

As well, the original error messages from the utils/actor_path.py
script has been print on stdout which had several negative effects:
  - the error messages has not been printed anywhere
  - error messages has been executed in the Makefile as a shell cmd
  - printed errors contained just the first word of the error msgs

This is kind of because the script in the Makefile ignores the exit
code of the utils script. So I have moved the error messages into
the stderr to print them always + added print of a special string
to stdout to ensure the execution of tests will be stopped
as expected (because the file names like the err msg will not exist
the most probably).