Add python3 workaround for el8toel9
See the description of the actor for more details. In short, the
/usr/bin/python3 is removed during the upgrade transaction (managed
by alternatives). The proper fix will be delivered in RHEL 8 & RHEL 9
packages. Until it's available, just create the alternatives if do
not exist after the RPM upgrade.

Skipping creation of unit tests as we expect to remove this actor
in couple of weeks.