Add the ScanFilesForTargetUserspace actor.
This actor allows us to specify what files should be copied to the target
userspace container upon its creation. Produces a `TargetUserSpacePreupgradeTasks`
containing the files to be copied.

The functionality provided by this actor is already employed to copy the
`/etc/hosts/` file into the target userspace, as the original implementation
relying on the usage of bind mounting would crash if the file did not exist.
The copy of this file is performed conditionally, that means it is copyied only
when the file exists. The original functionality that enables us to always bind
mount a certain file has been kept and it is still available for use.