Major packaging changes: leapp-repository renaming and conflicts
The leapp-repository rpm is going to be renamed based on the system
it is installed on:
  - RHEL 7  -> leapp-upgrade-el7toel8
  - RHEL 8  -> leapp-upgrade-el8toel9

The original naming has not been clear to people and we wanted
provide clearly what is the purpose of the RPM. The new naming
schema is:
    leapp-{purpose}-{main_reponame}

In this case, purpose of our repositories is now just IPU between
major versions of RHEL (so the leapp-upgrade part). For IPU 7 > 8 the
main leapp repository is el7toel8, ...

((As well this naming syncs our project with convert2rhel (hi guys!)
where the `leapp-convert` name will be used in future when they
switch to leapp.))

Additionally set conflicts between
    leapp-upgrade-el7toel8
    leapp-upgrade-el8toel9
packages to ensure they are not replaced during the upgrade as we
are super interested about keeping the installed leapp rpms untouched
during the upgrade process (excluding deps metapackages) to be sure
we finish the upgrade with the same RPMs as we started with.

Additionally use the `setup` macro instead of `autosetup`. We do not
expect to use any patches in upstream and the automatic application
was kind of fragile in downstream. So let's keep the `setup`
everywhere to be more synced.