Inhibit the upgrade when multiple rescue boot entries used on s390x.
When performing an IPU on s390x the upgrade will not finish
successfully, due to the failure of conversion of the Zipl configuration
to the BLS, when multiple Zipl rescue entries are present in the
configuration.

To address this problem the BootEntriesScanner was introduced along with
the `BootEntry` model, capable of scanning the bootloader configuration
and providing the information to other actors.

The added `ZiplCheckMultipleRescueEntries` actor consumes this
information and inhibits the upgrade if there are multiple rescue
entries found, and therefore, the upgrade would not be successful.
The actor runs only on the s390x architecture as multiple rescue entries
pose no problem when running the GRUB2 bootloader.