Packaging: rename the deps packages
The original name "leapp-repository-deps" has been changed
to correspond with the new naming:
    leapp-upgrade-el7toel8-deps
    leapp-upgrade-el8toel9-deps
So we are pretty sure which packages we have installed on the
system. Deps packages installed during the IPU (for the target
system) supposed to replace those for the source system, reflect
the change as well. But regarding the their purpose I've kept the
name we are using since the start.