Preserve leapp packages during the IPU
For the IPU process it's crucial to protect leapp and leapp-repository
packages (all but the deps packages, which are expected to be replaced
during the IPU). It's super important the packages are not replaced
or removed during the IPU. With the IPU 8 -> 9 we have to introduce
leapp packages in RHEL 8 as well, which naturally means that our
packages are updated if we do not care about them.

To handle this problem, we exclude all leapp packages from any DNF
transaction. In this case literally. The DNF is going to be configured
to exclude our packages from any operation and it will be expected
to change the DNF configuration after the upgrade manually to be
able to remove our packages after the IPU.

In short, we do it by adding our packages into the 'exclude' option
in the /etc/dnf/dnf.conf file. This happens in the target userspace
container and later on host as well so the change is persistent
(unless someone pass even these options).

So when any custom actor call `dnf update`, our packages will stay
untouched even after the upgrade.

The functionality is implemented in the dnfconfig.py library
which uses the `dnf config-manager` command. The command requires
installed specific dnf subpackage and it is not available on RHEL 7,
so all actions could be done only on RHEL 8+ (including the target
userspace container).