FIX: Add missing SYSTEMD_SECCOMP environment variable
During the improvements for RHEL9 compatibility we added the
SYSTEMD_SECCOMP environment variable for calls to DNF. This
positively impacts DNS resolution.
In case of the installation of additional packages for the initramfs
we did overlook this and it resulted in DNF not being able to resolve
the hostnames.
This patchs adds the SYSTEMD_SECCOMP environment variable also for the
DNF call do install additional packages.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>