Add detection of enabled yum plugins.
YUM provides the user with a functionality allowing him to enable/disable plugins
that are crucial for the IPU - `subscription-manager` and `product-id`.
Disabling those plugins causes the upgrade to fail, due to being unable
to obtain repositories that will be used on the target system.

In order to detect such system state, the `YumConfigScanner` actor was
introduced that allows to detect what plugins are being loaded. This is
done by calling yum without any parameters (except `-v` on RHEL8+) and
parsing the list of loaded plugins from the output. The parsing is done
in a fashion that accounts for possible line breaks in the YUM output.

Furthermore, a new check is introduced - the `CheckYumPluginsEnabled`
actor, that consumes the `YumConfig` and inhibits the upgrade if any of
the required plugins is not being loaded.