Refactoring - utilize get_target_major_version where appropriate.
Introduces changes required in the code review for the PR668. Majority
of changes related to the actual code lies in refactoring the code to
use the `get_target_major_version` function introduced in this PR where
appropriate, instead of relying on direct version string manipulation.

A new helper function `get_version_major` was introduced to encapsulate
the logic required to parse out the major version from a version string
and to provide a single spot where such manipulation is performed should
some additional logic be performed.

Additional changes include fixes to typos found in docstrings.