Enforce locale when running the IPU workflow.
Some actors that are executed during the IPU workflow rely on parsing
outputs of executed commands. When the underlying system is configured
with a different locale the actors typically fail since they are looking
for specific strings in the outputs which tend to be localized.

This commit enforces locale to en_US.utf-8 via setting the LANG
environment variable, and therefore removes the need for users that
typically use different locale to set it before running leapp.