rpmscanner: fix loading of data about module streams with DNF
The original solution worked just in case no DNF plugins habe been
required to load available repositories metadata. However e.g. in
case of RHUI, we need plugins to get such a data as baseurls of
repositories usually contain REGION substring that is expected to
be replaced by the Amazon-id plugin.

Added initilisation and configuration of DNF plugins as expected.