Disable pylint warnings about missing open() encoding parameter.
The new version of pylint adds a new warning that warns about not
specifying encoding when opening a file. This option was introduced
mostly to support better compatibility when running Python packages on
Windows, where the usage of the implicit encoding causes trouble.

Also adding two stylistic options to the disabled ones: use-dict-literal
and use-list-literal, both suggesting to use []/{} instead of calling
list() or dict().