Replace systemd-mount output parsing with pyudev.
StorageScanner relied on collecting information provided by the
`systemd-mount` command by parsing its output. The output format of the
command has proved to be error prone to parse due to its tabular nature
(a table padded with whitespace).

This commit changes the source of the information provided. Instead of
relying on calling the command and parsing its outputs, we query udev
directly using the pyudev library, closely mimicking what
`systemd-mount` does behind the scene. This way we are no longer parsing
commands output, eliminating a source of errors.