targetuserspacecreator: Improve the error msg when custom repos missing
We have found that people has troubles to realize where the problem
exactly is when they require specific repositories using the
--enablerepo option.

e.g. sometimes they do not have access to a Satellite server so they
cannot check what repositories are provided for the target system
and they do not know we print the list of available repositories
in the logs.

Additionally, it's taking a time to find which repositories are missing
or which contain a typo. So we print the list of missing repositories
to make this clean on what people should focus.

As the list of all available target repositories could be pretty long,
we do not print the list in the error message, but point people they
can find the list in the logs, so they are aware they have this
information available from us.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2002347