read_or_fetch: Add option allowing to specify expected encoding.
The requests library relies on Content-Encoding header to determine what
the encoding of the response is. As the response is compressed, the
actual encoding of the received data is unknown, so it relies on
heuristics to determine the actual encoding. This patch prevents
the library from guessing and decodes the response contents manually.

The read_or_fetch library now allows to specify what is the expected
encosing of the file that should be read. Default value is 'utf-8'.