Introduce RepositoriesMapping model and drop RepositoriesMap model
The new model is going to replace the original RepositoriesMap model
that has been used for the old repomap format data. As we are going
to switch into the new format, we are not able to use the old
models as it's incompatible with other use-cases we want to cover
with it.

The new model reflects data structure used in the input data file,
however it's already simplified as some pieces of data are needed
just for the scanner actor (e.g. the upgrade path is known to all
actors and doesn't have to be anyhow duplicated inside the repomap
data models).

Also we drop the original RepositoriesMap and RepositoryMap models
because:
  * they are not used anymore by any actors (as these have been
    replaced by new models)
  * we cannot deprecate these models as from the nature of the change,
    we cannot keep the functionality around nor produce the original
    message as data in such a msg could not provide the same meaning.
    That means, we cannot say strictly that repoidX will be mapped
    to repoidY as the logic of mapping is going to be changed
    completely.

Adding docstrings to inform people the new models are not supposed to
be consumed by other (custom) actors as most likely we drop them
in close future as well.