peseventsscanner: adapt to the new repomap msgs and add decision logic
With the new repomap data we have have to apply the mapping algoritmus
in each actor working with the repomap data that need to know target
repositories or pes ids. We do not want to create a shared library
to work with the repomap data so we put a private repomap library
into each such an actor. We expect to update all such an actors soon
(e.g. merge them together) to reduce the complexity of all these
actors.

- do not lower pes-ids (we lowered them (originally) because of different
  upper/lower chars used for pesids in pes data and repomap data
  but now we generate all PES IDs in the repomap based on the real
  PES ids. Let's use the same so when we debug any problems with
  data, we do not have to search for "SpecialIDs" when logged some
  "specialids".

Co-authored-by: MichalHe <mhecko@redhat.com>