setuptargetrepos: adapt to the new repomap msgs and introduce new decision logic
This is analogy to the previous commit. In this case we update the
actor responsible for the 'final' mapping to the target repositories.
Actor has been refactored completely to make it more readable
with several bugfixes (e.g. removal of death code, ..).

The original decision logic for repomapping was this:
   1. each src repoid map to ekvivalent target repoid + ga repoid
      (basically to each target repoid mentioned in the repomap data)
   2. blacklisted drop blacklisted repositories
   3. add all required custom target repositories

IOW, it was typical that systems with EUS repositories expected EUS+GA
repositories:
    src repos       || dst repos
    ----------------||---------------------
    EUSbase         || EUSbaseos, GAbaseos, EUSappstream, GAappstream
    EUSbase, GAbase || EUSbaseos, GAbaseos, EUSappstream, GAappstream
    GAbase          || GAbase, GAappstream
    ...

The new logic works with families of repositories (based on pesids),
so we decide to pick just the best fitting repository from each mapped
family. So the table above is going to be changed like this:
    src repos       || dst repos
    ----------------||---------------------
    EUSbase         || EUSbaseos, EUSappstream
    EUSbase, GAbase || EUSbaseos, EUSappstream
    GAbase          || GAbase, GAappstream
    ...
Each target repo family (pesid) has just one best candidate
repository (or nothing in case of issues in mapping data).
The priority of repositories is
    beta < ga < all_else
where all else is e.g. eus, tus, aus, ... as we do not expect to have
multiple premium channels available & enabled on one machine.
In case of mapping between repositories have very special channels
which does not follow typical channels (e.g. extras is just one, doesn't
matter whether you have eus, aus, ga, ...it's same everywhere) or in
case the target repo family doesn't contain a special channel (yet?)
there is always fallback to default channels. In this case, we consider
default channels to be get based on the top prio repository from base
/ baseos families and always the 'ga'. So e.g. in case an EUS repository
is mapped to a family that has just a repo with the 'ga' channel,
this one repository is going to be used for the fallback.

Co-authored-by: MichalHe <mhecko@redhat.com>