repositoriesblacklist: update the actor to use new repomap msg
The original actor tried to blacklist target repositories based on
present Optional repositories (in case no Optional repo has been
enabled). The originally consummed msg is not produced anymore and
regarding the new concept of the repository mapping, we decided to
blacklist all CRB repositories - even those that could be irrelevant
for the particular system (e.g. RHUI CRB repoids on non-rhui systems,
etc). This simplifies the solution significantly and should not be
problematic for people. The most probably the solution will be updated
in future to report just relevant CRB repositories. But it's not
our current plan.

Co-authored-by: MichalHe <mhecko@redhat.com>