Introduce fast_lint make target.
Currently the makefile contains the 'lint' target that runs the linter
on the entire repository which takes a considerable amount of time. This
commit introduces 'fast_lint' target that runs the linter only on files
that have changed on the current branch compared to the master branch.
Moreover, the linter is run in parallel to provide even shorter feedback
time for the developer.