devel: Add persistent package cache feature
This feature will allow to reduce the time for multiple executions of
`leapp upgrade` or `leapp preupgrade` by reusing already downloaded
packages. This feature is meant to be used only for the part before the
reboot.

The feature can be activated by setting the following environment
variable:
```
LEAPP_DEVEL_USE_PERSISTENT_PACKAGE_CACHE=1
```
The value has to be 1 as it is checking against that explicitly.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>