targetuserspacecreator: handle properly yum.repos.d preparation for RHUI
When creating the target userspace container, we have to handle
the /etc/yum.repos.d/ directory inside the container to provide
all required repofiles we expect to have inside the container.
Usually we just copied the original /etc/yum.repos.d from the
scratch container (and rm -rf the dir inside the container before
that).

However, we have problem especially in case of RHEL on RHUI. In such
cases the repofiles for the target RHEL systems are delivered via
the el8 rhui rpms which are installed inside the container. When
we use just an original repofiles we used to create the container,
the download of rpms fails with msgs about invalid certificates,
etc. As well, we cannot keep the old rhui repofiles present as they
could lead into duplicated data about repositories (repoids...) and
will lead to very similar issues.

To handle this correctly without breaking other solutions, the target
yum.repos.d directory has to contain:
  - "original" repofiles - excluding any rhui repofiles (leapp + these
    owned by RHUI [client] rpms) - and
  - repofiles owned by any rpms installed inside the target userspace
    container

The current solution is written in kinda hacky way, but we are going
to live with that until the planned refactoring of the actor (which
should happen relatively soon already).