Updated the deprecated dates for the newly deprecated entities
As the upcoming release is after a long time since the last one,
we want to update dates in the deprecation messages so people do not
see 'high risk' reports first day after the release. Regarding we
want to keep stuff working 6months after the release that introduces
new deprecated stuff it's better to update all deprecated. We do not
need to put there any "future" date as we are ok if people see
something more highlighted a month before the end of protection period.