Fix the rerun command regarding the --channel ....
With the introduction of the --channel option for the leapp upgrade
command, the rerun command has become broken as it doesn't know
anything about this option but part of the code is shared with the
upgrade command. Introduce the channel field in the `Namespace`
object that represents the "args" in case of the rerun command.