unittests: Scenario based testing
Previously we tested all repositories all together. We didn't really
care wether it was for el7toel8 or el8toel9, we even had el8toel9
being tested with python2.7 python until we deleted the el8toel9
repository within the GitHub action as a temporary measure.

This implements the feature of allowing el7toel8 and el8toel9 testing
scenarios to be implemented properly.
This means as of this patch there are 4 unit test runs:
- Python 2.7 with system_upgrade/el7toel8 and system_upgrade/common
  repos in a CentOS7 container
- Python 3.6 with system_upgrade/el7toel8 and system_upgrade/common
  repos in a CentOS8 container
- Python 3.6 with system_upgrade/el8toel9 and system_upgrade/common
  repos in a CentOS8 container
- Python 3.9 with system_upgrade/el8toel9 and system_upgrade/common
  repos in a CentOS8 container

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>