transactionworkarounds: Refactoring and fixed tests
Previously the transactionworkarounds actor was only executed and did
not verify the actual work it does, but instead ran it and assumed it
will get the bundled RPM packages from the el7toel8 repo. As of this
moment in the el8toel9 scenario there are no bundled rpms, the test
fails if only the system_upgrade/el8toel9 is loaded but not the
system_upgrade/el7toel8 repo. This showed that the test is flawed and
needed to be rewritten.
To properly test the actor, the code had to be moved into a library so
it can be properly mocked for the unit tests.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>