Makefiles: Use conditional installs to avoid duplications
Previously some of the packages in the Makefiles were installed with
ignoring the exit code. In case there were the either or condition, the
second package might have been installed as well, even though it wasn't
necessary. This patch changes this by making it conditional, to install
the second package only if the first one failed.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>