Makefile: fix getting paths of leapp repos for unit-tests
Previously on some systems the generation of SYSUPG_TEST_PATHS
has been broken. Instead of string like:
  "path/repo1 path/repod2 ..."
it generated string like:
  "path/repo1path/repod2..."
So it return non-existent paths. Strangely, this happened just on
some machines with the same installed system. Not sure what is the
difference, but this implementation is more generic and should work
without any magic.