Do not crash when the next EFI boot cannot be set
We want to set the EFI next-boot entry to be the current booted one
as the default boot entry could be a special "handler" for machines.
It could be used e.g. on machines that are part of testing
infrastructure, to service them (e.g. put them back to the pool of
available machines, ...).

However, for IPU 8 -> 9 we have reported error on Hyper-V, where we
are trying to set the next boot to a record that doesn't exist
anymore. Which is weird as we obtain the record from the efibootmgr
itself just before we try to set it. Possibly there is a bug in the
efibootmgr, not sure. For now, print warning when we cannot set the
next boot with additional information available, but do not crash
in such a case.

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=2024059