modularity: Reset modules before upgrade
Previously we have added the modularity support into leapp. However we
struggled with the part of resetting the modules. This patch will add
the missing piece for that feature.
It implements resetting modules for module streams that are not going to
be installed. If the same module gets installed, but a different stream,
the module needs to be reset as well. The only exception is when the
module:stream tuple is the same, then we should rather not reset it,
according to the DNF developers.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>