Zipl: Inhibit if boot entries share the same kernel image
This patch extends the detection of invalid zipl configuration
(introduced in PR700) by detecting also boot entries that share the same
kernel image version. Such entries cause problems when converting
the boot configuration to BLS using the zipl-switch-to-blscfg script.
The detection logic has been developed in such a way that it mimics the
internal logic of this script, and thus, it covers all scenarios in
which the boot loader configuration might cause the upgrade to fail.