Provide LEAPP_IPU_IN_PROGRESS envar
The environement variable contains string in format XtoY, where
  X is major version of the current system
  Y is major version of the target system

So contains:
  7to8 for IPU 7 -> 8,
  8to9 for IPU 8 -> 9

this envar is supposed to be used e.g. in RPM scriptlets in case
maintainers want to treat the upgrade correctly during the IPU
process between major releases of RHEL.

As a benefit, this allows people to handle situations when specific
action has to happen in the middle of the DNF transaction - as leapp
actors can do actions only before and after the transaction.